move fullscreen-button component to common folder
