+++ bigbluebutton-html5/imports/ui/components/common/fullscreen-button/container.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import FullscreenButtonComponent from './component';
-import { layoutSelect, layoutDispatch } from '../layout/context';
+import { layoutSelect, layoutDispatch } from '/imports/ui/components/layout/context';
 import FullscreenService from './service';
 
 const FullscreenButtonContainer = (props) => <FullscreenButtonComponent {...props} />;
