+++ bigbluebutton-html5/imports/ui/components/common/fullscreen-button/service.js
