+++ bigbluebutton-html5/imports/ui/components/tooltip/container.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { withTracker } from 'meteor/react-meteor-data';
-import FullscreenService from '/imports/ui/components/fullscreen-button/service';
+import FullscreenService from '/imports/ui/components/common/fullscreen-button/service';
 import Tooltip from './component';
 
 const TooltipContainer = props => <Tooltip {...props} />;
