+++ bigbluebutton-html5/imports/ui/components/video-provider/video-list/video-list-item/component.jsx
@@ -4,8 +4,8 @@ import browserInfo from '/imports/utils/browserInfo';
 import { Meteor } from 'meteor/meteor';
 import PropTypes from 'prop-types';
 import BBBMenu from '/imports/ui/components/menu/component';
-import FullscreenService from '/imports/ui/components/fullscreen-button/service';
-import FullscreenButtonContainer from '/imports/ui/components/fullscreen-button/container';
+import FullscreenService from '/imports/ui/components/common/fullscreen-button/service';
+import FullscreenButtonContainer from '/imports/ui/components/common/fullscreen-button/container';
 import Styled from './styles';
 import VideoService from '../../service';
 import {
